+++ services/notification/mail/mail.go
@@ -12,8 +12,8 @@ import (
 	repo_model "code.gitea.io/gitea/models/repo"
 	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/modules/log"
-	"code.gitea.io/gitea/modules/notification/base"
 	"code.gitea.io/gitea/services/mailer"
+	"code.gitea.io/gitea/services/notification/base"
 )
 
 type mailNotifier struct {
